# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
    "version": "1.0.0",
    "title": "dataProduct API",
    "contact": {
      "email": "jay.thesia.23@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http"],
  "host": "localhost:5050",
  "basePath": "/api",
  "paths": {
    "/productshow": {
      "get": {
        "summary": "Show the data product",
        "description": "Show the data product",
        "products": ["/application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductResponse"
              }
            }
          },
          "400": {
            "description": "failed opertion",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },

    "/productadd": {
      "post": {
        "summary": "Save the data product",
        "description": "Save the data product",
        "products": ["/application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "product": {
                  "type": "object",
                  "$ref": "#/definitions/Product"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },

    "/productdelete/{id}": {
      "delete": {
        "description": "Delete the products",
        "products": "[application/json]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductResponse"
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },

    "/productupdate/{id}": {
      "put": {
        "summary": "Update the data product",
        "description": "update the product",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "which id data should be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": "true",
            "schema": {
              "type": "object",
              "properties": {
                "product": {
                  "type": "object",
                  "$ref": "#/definitions/Product"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    }
  },

  "definitions": {
    "ProductResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "product": {
          "type": "string"
        }
      }
    },

    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "product": {
          "type": "string"
        }
      }
    },

    "InvalidResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}